merge in latest

+++ tools/util.py
@@ -6,20 +6,42 @@ import stat
 import sys
 import subprocess
 
+RESET = "\x1b[0m"
+FG_RED = "\x1b[31m"
+FG_GREEN = "\x1b[32m"
+
 executable_suffix = ".exe" if os.name == "nt" else ""
 root_path = os.path.dirname(os.path.dirname(os.path.realpath(__file__)))
 
 
-def make_env(merge_env={}, env=None):
+def make_env(merge_env=None, env=None):
     if env is None:
         env = os.environ
     env = env.copy()
+    if merge_env is None:
+        merge_env = {}
     for key in merge_env.keys():
         env[key] = merge_env[key]
     return env
 
 
-def run(args, quiet=False, cwd=None, env=None, merge_env={}):
+def add_env_path(add, env, key="PATH", prepend=False):
+    dirs_left = env[key].split(os.pathsep) if key in env else []
+    dirs_right = add.split(os.pathsep) if isinstance(add, str) else add
+
+    if prepend:
+        dirs_left, dirs_right = dirs_right, dirs_left
+
+    for d in dirs_right:
+        if not d in dirs_left:
+            dirs_left += [d]
+
+    env[key] = os.pathsep.join(dirs_left)
+
+
+def run(args, quiet=False, cwd=None, env=None, merge_env=None):
+    if merge_env is None:
+        merge_env = {}
     args[0] = os.path.normpath(args[0])
     if not quiet:
         print " ".join(args)
@@ -30,7 +52,9 @@ def run(args, quiet=False, cwd=None, env=None, merge_env={}):
         sys.exit(rc)
 
 
-def run_output(args, quiet=False, cwd=None, env=None, merge_env={}):
+def run_output(args, quiet=False, cwd=None, env=None, merge_env=None):
+    if merge_env is None:
+        merge_env = {}
     args[0] = os.path.normpath(args[0])
     if not quiet:
         print " ".join(args)
@@ -59,14 +83,24 @@ def shell_quote(arg):
         return quote(arg)
 
 
+def red_failed():
+    return "%sFAILED%s" % (FG_RED, RESET)
+
+
+def green_ok():
+    return "%sok%s" % (FG_GREEN, RESET)
+
+
 def remove_and_symlink(target, name, target_is_dir=False):
+    if os.name != "nt" and os.path.islink(name):
+        return
     try:
         # On Windows, directory symlink can only be removed with rmdir().
         if os.name == "nt" and os.path.isdir(name):
             os.rmdir(name)
         else:
             os.unlink(name)
-    except:
+    except OSError:
         pass
     symlink(target, name, target_is_dir)
 
@@ -117,15 +151,26 @@ def touch(fname):
 
 
 # Recursive search for files of certain extensions.
-# (Recursive glob doesn't exist in python 2.7.)
-def find_exts(directory, *extensions):
+#   * Recursive glob doesn't exist in python 2.7.
+#   * On windows, `os.walk()` unconditionally follows symlinks.
+#     The `skip`  parameter should be used to avoid recursing through those.
+def find_exts(directories, extensions, skip=None):
+    if skip is None:
+        skip = []
+    assert isinstance(directories, list)
+    assert isinstance(extensions, list)
+    skip = [os.path.normpath(i) for i in skip]
     matches = []
-    for root, dirnames, filenames in os.walk(directory):
-        for filename in filenames:
-            for ext in extensions:
-                if filename.endswith(ext):
-                    matches.append(os.path.join(root, filename))
-                    break
+    for directory in directories:
+        for root, dirnames, filenames in os.walk(directory):
+            if root in skip:
+                dirnames[:] = []  # Don't recurse further into this directory.
+                continue
+            for filename in filenames:
+                for ext in extensions:
+                    if filename.endswith(ext):
+                        matches.append(os.path.join(root, filename))
+                        break
     return matches
 
 
@@ -147,12 +192,12 @@ def build_mode(default="debug"):
         return default
 
 
-# E.G. "out/debug"
+# E.G. "target/debug"
 def build_path():
     if "DENO_BUILD_PATH" in os.environ:
         return os.environ["DENO_BUILD_PATH"]
     else:
-        return os.path.join(root_path, "out", build_mode())
+        return os.path.join(root_path, "target", build_mode())
 
 
 # Returns True if the expected matches the actual output, allowing variation
@@ -212,7 +257,7 @@ def enable_ansi_colors_win10():
 
     # Function factory for errcheck callbacks that raise WinError on failure.
     def raise_if(error_result):
-        def check(result, func, args):
+        def check(result, _func, args):
             if result == error_result:
                 raise ctypes.WinError(ctypes.get_last_error())
             return args
@@ -286,7 +331,7 @@ def enable_ansi_colors_win10():
     # Try to set the flag that controls ANSI escape code support.
     try:
         SetConsoleMode(conout, mode.value | ENABLE_VIRTUAL_TERMINAL_PROCESSING)
-    except WindowsError as e:
+    except WindowsError as e:  # pylint:disable=undefined-variable
         if e.winerror == ERROR_INVALID_PARAMETER:
             return False  # Not supported, likely an older version of Windows.
         raise
@@ -297,7 +342,6 @@ def enable_ansi_colors_win10():
 
 
 def parse_unit_test_output(output, print_to_stdout):
-    first = True
     expected = None
     actual = None
     result = None
